+++ filament/src/View.cpp
@@ -677,7 +677,7 @@ void FView::prepareStructure(backend::Handle<backend::HwTexture> structure) cons
 }
 
 void FView::prepareShadow(backend::Handle<backend::HwTexture> texture) const noexcept {
-    mShadowMapManager.prepareShadow(texture, mPerViewSb);
+    mShadowMapManager.prepareShadow(texture, *this);
 }
 
 void FView::cleanupRenderPasses() const noexcept {
@@ -993,6 +993,18 @@ void View::setDynamicLightingOptions(float zLightNear, float zLightFar) noexcept
     upcast(this)->setDynamicLightingOptions(zLightNear, zLightFar);
 }
 
+void View::setShadowType(View::ShadowType shadow) noexcept {
+    upcast(this)->setShadowType(shadow);
+}
+
+void View::setVsmShadowOptions(VsmShadowOptions const& options) noexcept {
+    upcast(this)->setVsmShadowOptions(options);
+}
+
+View::VsmShadowOptions View::getVsmShadowOptions() const noexcept {
+    return upcast(this)->getVsmShadowOptions();
+}
+
 void View::setAmbientOcclusion(View::AmbientOcclusion ambientOcclusion) noexcept {
     upcast(this)->setAmbientOcclusion(ambientOcclusion);
 }
@@ -1003,10 +1015,6 @@ View::AmbientOcclusion View::getAmbientOcclusion() const noexcept {
 
 void View::setAmbientOcclusionOptions(View::AmbientOcclusionOptions const& options) noexcept {
     upcast(this)->setAmbientOcclusionOptions(options);
-}
-
-void View::setShadowType(View::ShadowType shadow) noexcept {
-    upcast(this)->setShadowType(shadow);
 }
 
 View::AmbientOcclusionOptions const& View::getAmbientOcclusionOptions() const noexcept {