+++ libs/viewer/src/Settings.cpp
@@ -192,6 +192,27 @@ static int parse(jsmntok_t const* tokens, int i, const char* jsonChunk, ShadowTy
 }
 
 static int parse(jsmntok_t const* tokens, int i, const char* jsonChunk,
+        VsmShadowOptions* out) {
+    CHECK_TOKTYPE(tokens[i], JSMN_OBJECT);
+    int size = tokens[i++].size;
+    for (int j = 0; j < size; ++j) {
+        const jsmntok_t tok = tokens[i];
+        CHECK_KEY(tok);
+        if (0 == compare(tok, jsonChunk, "anisotropy")) {
+            i = parse(tokens, i + 1, jsonChunk, &out->anisotropy);
+        } else {
+            slog.w << "Invalid shadow options key: '" << STR(tok, jsonChunk) << "'" << io::endl;
+            i = parse(tokens, i + 1);
+        }
+        if (i < 0) {
+            slog.e << "Invalid shadow options value: '" << STR(tok, jsonChunk) << "'" << io::endl;
+            return i;
+        }
+    }
+    return i;
+}
+
+static int parse(jsmntok_t const* tokens, int i, const char* jsonChunk,
         TemporalAntiAliasingOptions* out) {
     CHECK_TOKTYPE(tokens[i], JSMN_OBJECT);
     int size = tokens[i++].size;
@@ -555,6 +576,8 @@ static int parse(jsmntok_t const* tokens, int i, const char* jsonChunk, ViewSett
             i = parse(tokens, i + 1, jsonChunk, &out->dynamicLighting);
         } else if (compare(tok, jsonChunk, "shadowType") == 0) {
             i = parse(tokens, i + 1, jsonChunk, &out->shadowType);
+        } else if (compare(tok, jsonChunk, "vsmShadowOptions") == 0) {
+            i = parse(tokens, i + 1, jsonChunk, &out->vsmShadowOptions);
         } else if (compare(tok, jsonChunk, "postProcessingEnabled") == 0) {
             i = parse(tokens, i + 1, jsonChunk, &out->postProcessingEnabled);
         } else {
@@ -686,6 +709,7 @@ void applySettings(const ViewSettings& settings, View* dest) {
     dest->setDynamicLightingOptions(settings.dynamicLighting.zLightNear,
             settings.dynamicLighting.zLightFar);
     dest->setShadowType(settings.shadowType);
+    dest->setVsmShadowOptions(settings.vsmShadowOptions);
     dest->setPostProcessingEnabled(settings.postProcessingEnabled);
 }
 
@@ -1007,6 +1031,14 @@ std::string writeJson(const DynamicLightingSettings& in) {
     return oss.str();
 }
 
+std::string writeJson(const VsmShadowOptions& in) {
+    std::ostringstream oss;
+    oss << "{\n"
+        << "\"anisotropy\": " << writeJson(in.anisotropy) << "\n"
+        << "}";
+    return oss.str();
+}
+
 std::string writeJson(const ViewSettings& in) {
     std::ostringstream oss;
     oss << "{\n"
@@ -1023,6 +1055,7 @@ std::string writeJson(const ViewSettings& in) {
         << "\"renderQuality\": " << writeJson(in.renderQuality) << ",\n"
         << "\"dynamicLighting\": " << writeJson(in.dynamicLighting) << ",\n"
         << "\"shadowType\": " << writeJson(in.shadowType) << ",\n"
+        << "\"vsmShadowOptions\": " << writeJson(in.vsmShadowOptions) << ",\n"
         << "\"postProcessingEnabled\": " << writeJson(in.postProcessingEnabled) << "\n"
         << "}";
     return oss.str();