+++ RELEASE_NOTES.md
@@ -5,6 +5,9 @@ A new header is inserted each time a *tag* is created.
 
 ## Next release (main branch)
 
+- Added View::setVsmShadowOptions (experimental)
+- Add anisotropic shadow map sampling with VSM (experimental)
+
 ## v1.9.5
 
 - Added a new Live Wallpaper Android sample